refactor(lint): hoist shared eslint + prettier config to root
