+++ apps/shared/eslint.config.mjs
@@ -1,0 +1,5 @@
+import shared from '../../eslint.config.shared.mjs'
+
+export default [
+  ...shared
+]
