+++ apps/shared/package.json
@@ -8,10 +8,16 @@
   },
   "types": "./src/index.ts",
   "scripts": {
+    "lint": "eslint src/",
+    "lint:fix": "eslint src/ --fix",
     "typecheck": "tsc -p . --noEmit",
     "check": "npm run typecheck"
   },
   "devDependencies": {
-    "typescript": "^6.0.3"
+    "@eslint/js": "^9.39.4",
+    "eslint": "^9.39.4",
+    "globals": "^17.4.0",
+    "typescript": "^6.0.3",
+    "typescript-eslint": "^8.56.1"
   }
 }
